fix: byte within binaryType
