+++ src/types/binary/index.ts
@@ -1,3 +1,3 @@
 export { Word } from './word'
 export { Halfword } from './halfword'
-export { Byte } from './byte'
+export { Byte } from './binaryType'
